Update App.js
Did some cleaning.

( line 14) In a loose equality ( == ) null=undefined and vice verse since a loose equality doesn't check for the type of value. In a strict equality ( === ) null=/=undefined since they are 2 different types. Undefined is a type in itself while null is an object. Check console.log( typeof(null) )

Shorthand on line 86 doesn't need to be on multiple lines.